pyproject.toml
[project]
name = "venue-intel"
version = "0.1.0"
description = "Venue Intelligence & Distribution Prioritisation System"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31.0",
    "pydantic>=2.5.0",
    "pandas>=2.1.0",
    "openpyxl>=3.1.0",      # Excel export
    "python-dotenv>=1.0.0", # Environment variables
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
    "pytest>=7.4.0",
    "ipykernel>=6.0.0",     # Jupyter support
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
]

[tool.ruff.lint.isort]
known-first-party = ["venue_intel"]
